Extend the HDMI info to extract picture aspect ratio from kernel modes.
BZ: 49783

It is necessary to inspect and extract picture aspect ratio from the
display modes read from kernel. This is required to have the ability
to accurately set the particular mode (eg: vic 18: 720x576p@50 [16:9]
instead of vic 17: 720x576p@50 [4:3]) for compliance testing.

All mode match comparisons, now need width, height, refresh rate and
relevant flags (interlace and aspect ratio) to be compared prior to
deciding a match found.

Made the relevant extensions to the interface which is used by the
HDMI Settings app to display this information.

Change-Id: I115db4b3607b796b00ebdba56a2991dc67020369
Signed-off-by: Arun Kannan <arun.kannan@intel.com>
Reviewed-on: http://android.intel.com:8080/61920
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>