Ensure video refresh rate request in HDMI extended mode is honoured.
BZ: 54100

While trying to look for a matching mode in video extended mode, the
order for selecting a mode is:
1. Find a completely matching mode (width, height, refresh, flags).
2. Find a mode which at least matches video refresh rate.

The condition for at least matching the refresh was failing due to
"flags" field comparison which is not required for condition '2'. It
is also necessary for the video case, to set the ratio field as "0"
to make sure it is treated as "unknown" for purpose of mode selection.

If the given video matches condition '1', we go with that mode. If
not, we go with the display mode that at least matches the refresh rate
as given in condition '2'. If both dont find a match, we simply use the
HDMI display's preferred display mode.

Change-Id: I443666c61ebd61c6ea4d795938c5662ef2f1800d
Signed-off-by: Arun Kannan <arun.kannan@intel.com>
Reviewed-on: http://android.intel.com:8080/63625
Reviewed-by: Zhu, Tianyang <tianyang.zhu@intel.com>
Reviewed-by: Liu, Wei W <wei.w.liu@intel.com>
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>