[Video-HDMI]: Prevent Incall end screen shown on HDMI monitor/TV.
BZ: 37432 61057

This patch is ported from R3 based on the patch 57612.

Currently, display observer is relying on the
com.android.phone_INCALLSCREEN_FINISH intent which is
broadcast by the phone application when the InCallScreen
is ended to display the screen on HDMI device. This is
not working all the time. Also as per Android cdd, no new
intent should be introduced in android.* and com.android.*

With this patch, instead of listening for
com.android.phone_INCALLSCREEN_FINISH intent, current
running task is checked to determine whether to
allow HDMI ON or not. If the InCallScreen is the foreground
task, then a timer of 3seconds is configured for recheck.
If the InCallScreen is not active, then HDMI on is allowed.

Becasue r3/ICS is hard to merge patch, and the bug 37432 had been moved
to R4, we will do a full and carefully
test on r4/JB,then decide whether to port this patch and 66531 to r3/ICS.

Change-Id: I373f108e17e4a650af5ce23016eb3ba3f505d188
Signed-off-by: Zhu, Tianyang <tianyang.zhu@intel.com>
Reviewed-on: http://android.intel.com:8080/70728
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Reviewed-by: Xie, Lin <lin.xie@intel.com>
Reviewed-by: Jia, Lin A <lin.a.jia@intel.com>
Reviewed-by: Wang, Lili A <lili.a.wang@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>