WiDi: [Multi-tasking] Verify if surface is created by MDS for play in background.
BZ: 64895

In Widi Multitasking mode, Surface is created by
MDS, at hardware composer we have surface information
that information is send to MDS to verify if MDS
is the owner of that surface.

This patch verifies if native window is from
the surface created by MDS during video playback
in background mode.

Change-Id: I8cecfa2be8b9373a2f05ed6d2d14b149476473e5
Signed-off-by: ashish singhi <ashish.singhi@intel.com>
Reviewed-on: http://android.intel.com:8080/72052
Reviewed-by: Zhu, Tianyang <tianyang.zhu@intel.com>
Reviewed-by: Sun, Hang L <hang.l.sun@intel.com>
Tested-by: Post, DavidX J <davidx.j.post@intel.com>
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Tested-by: lab_cactus <lab_cactus@intel.com>