HDMI: Refine MDS for BayTrail and MERR
BZ: 108143

This is the initial version after the MDS refine.

The changes are for the following targets:
1) Only keep the necessary duties in MDS. Since the Android MR1
framework supports the external display, some GFX DRM "write" work
are moved to hwcomposer. Such as turn on/off MIPI, hotplug detection and hdcp.

2) MDS serves video playback on HDMI/WiDi, such as broadcast
the video source info, video playback state, set hdmi timing.

3) MDS provides callback to clients for interaction. The listener
works the same as before. A new IMulitDisplayCallback is introduced
and it requires the module which has "write" permission of GFX DRM to implement
all the APIs and register the callback back to MDS. This callback
defines some new APIs to replaced the old "onMDSMessage". Now the task
is taken by hwcomposer.

4) The DisplayObserver in Java layer doesn't receive the hotplug
event any more. Some obsolete code is removed.

The names of some APIs and structure are changed. Because CTP and
Baytrail uses other versions of MDS, they should also be changed
for alignment.

Change-Id: I9eb5f62e9c57b2f5d4541c4f7e2300e9f07a21db
Signed-off-by: Tianyang Zhu <tianyang.zhu@intel.com>
Signed-off-by: Zhu,Tianyang <tianyang.zhu@intel.com>
Reviewed-on: http://android.intel.com:8080/105258
Reviewed-by: Xie, Lin <lin.xie@intel.com>
Reviewed-by: Wang, Lili A <lili.a.wang@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>