Video-HDMI: refine the scaling related APIs in MDS.
BZ:123432

The APIs for setting overscan&scaletype will check the implementation
from MDSCallback first.
If it is not implemented in the callback, MDS then will use the way by
SurfaceFlinger.

Note: As the dynamic mode setting is going to be moved to hwc, keeping
the extended mode state in MDS may not be needed.

Change-Id: Ie0530f6b100473e34005aacbdfd4305528ec264d
Signed-off-by: Lin Xie <lin.xie@intel.com>
Reviewed-on: http://android.intel.com:8080/118605
Reviewed-by: Wang, Lili A <lili.a.wang@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>