[Graphics]: fix the wrong type copy to cause HDMI setting no effect
BZ: 160632

The MDSHdmiTiming is copy wrong type in BnMultiDisplayHdmiControl.
This will cause the value of ratio is totally invalid.
So before Multidisplay send HDMIsetting to kernel, it will check this
HDMISetting in its list. Because the ratio value is totally invalid.
There is no proper HdmiTiming can be set. So it will directly return wrong.
And no mode setting request send to kernel.
This is the root cause of this issue.

Change-Id: Ia43d5415b63248d0af57532d36a5e3056c0042f2
Signed-off-by: Shi Yang <yang.a.shi@intel.com>